requirements.txt
appdirs==1.4.4
argcomplete==3.3.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Cython==3.0.10
diff-match-patch==20230430
Django==5.0.6
django-admin-logs==1.2.0
django-admin-rangefilter==0.13.0
django-colorfield==0.11.0
django-import-export==4.0.7
django-jquery==3.1.0
django-mathfilters==1.0.0
idna==3.7
import-export==0.3.1
importlib_resources==6.4.0
meson==1.4.1
meson-python==0.16.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pypng==0.20220715.0
pyproject-metadata==0.8.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-escpos==3.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
tenacity==8.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
versioneer==0.29
